requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
shapely
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio